feat: unify structure and add types 5.005, 6.010 and 20.102
