+++ knx/dpt/types_1_test.go
@@ -59,3 +59,129 @@ func TestDPT_1(t *testing.T) {
 		}
 	}
 }
+
+// Test DPT 1.001 (Switch) with values within range
+func TestDPT_1001(t *testing.T) {
+	var buf []byte
+	var src, dst DPT_1001
+
+	for _, value := range []bool{true, false} {
+		src = DPT_1001(value)
+		if bool(src) != value {
+			t.Errorf("Assignment of value \"%v\" failed! Has value \"%s\".", value, src)
+		}
+		buf = src.Pack()
+		dst.Unpack(buf)
+		if bool(dst) != value {
+			t.Errorf("Wrong value \"%s\" after pack/unpack! Original value was \"%v\".", dst, value)
+		}
+	}
+}
+
+// Test DPT 1.002 (Bool) with values within range
+func TestDPT_1002(t *testing.T) {
+	var buf []byte
+	var src, dst DPT_1002
+
+	for _, value := range []bool{true, false} {
+		src = DPT_1002(value)
+		if bool(src) != value {
+			t.Errorf("Assignment of value \"%v\" failed! Has value \"%s\".", value, src)
+		}
+		buf = src.Pack()
+		dst.Unpack(buf)
+		if bool(dst) != value {
+			t.Errorf("Wrong value \"%s\" after pack/unpack! Original value was \"%v\".", dst, value)
+		}
+	}
+}
+
+// Test DPT 1.003 (Enable) with values within range
+func TestDPT_1003(t *testing.T) {
+	var buf []byte
+	var src, dst DPT_1003
+
+	for _, value := range []bool{true, false} {
+		src = DPT_1003(value)
+		if bool(src) != value {
+			t.Errorf("Assignment of value \"%v\" failed! Has value \"%s\".", value, src)
+		}
+		buf = src.Pack()
+		dst.Unpack(buf)
+		if bool(dst) != value {
+			t.Errorf("Wrong value \"%s\" after pack/unpack! Original value was \"%v\".", dst, value)
+		}
+	}
+}
+
+// Test DPT 1.008 (Up/Down) with values within range
+func TestDPT_1008(t *testing.T) {
+	var buf []byte
+	var src, dst DPT_1008
+
+	for _, value := range []bool{true, false} {
+		src = DPT_1008(value)
+		if bool(src) != value {
+			t.Errorf("Assignment of value \"%v\" failed! Has value \"%s\".", value, src)
+		}
+		buf = src.Pack()
+		dst.Unpack(buf)
+		if bool(dst) != value {
+			t.Errorf("Wrong value \"%s\" after pack/unpack! Original value was \"%v\".", dst, value)
+		}
+	}
+}
+
+// Test DPT 1.009 (OpenClose) with values within range
+func TestDPT_1009(t *testing.T) {
+	var buf []byte
+	var src, dst DPT_1009
+
+	for _, value := range []bool{true, false} {
+		src = DPT_1009(value)
+		if bool(src) != value {
+			t.Errorf("Assignment of value \"%v\" failed! Has value \"%s\".", value, src)
+		}
+		buf = src.Pack()
+		dst.Unpack(buf)
+		if bool(dst) != value {
+			t.Errorf("Wrong value \"%s\" after pack/unpack! Original value was \"%v\".", dst, value)
+		}
+	}
+}
+
+// Test DPT 1.010 (Start) with values within range
+func TestDPT_1010(t *testing.T) {
+	var buf []byte
+	var src, dst DPT_1010
+
+	for _, value := range []bool{true, false} {
+		src = DPT_1010(value)
+		if bool(src) != value {
+			t.Errorf("Assignment of value \"%v\" failed! Has value \"%s\".", value, src)
+		}
+		buf = src.Pack()
+		dst.Unpack(buf)
+		if bool(dst) != value {
+			t.Errorf("Wrong value \"%s\" after pack/unpack! Original value was \"%v\".", dst, value)
+		}
+	}
+}
+
+// Test DPT 1.100 (Heat/Cool) with values within range
+func TestDPT_1100(t *testing.T) {
+	var buf []byte
+	var src, dst DPT_1100
+
+	for _, value := range []bool{true, false} {
+		src = DPT_1100(value)
+		if bool(src) != value {
+			t.Errorf("Assignment of value \"%v\" failed! Has value \"%s\".", value, src)
+		}
+		buf = src.Pack()
+		dst.Unpack(buf)
+		if bool(dst) != value {
+			t.Errorf("Wrong value \"%s\" after pack/unpack! Original value was \"%v\".", dst, value)
+		}
+	}
+}
